Github repository can be overridden in command line by Breeze
During testing v1-10-test backport for Breeze the
--github-repository flag did not work. It turned out that
the lowercase variable was not re-set when the flag was
provided by Breeze.

This change causes the lowercasing to be run just before it
is used to make sure that the GITHUB_REPOSITORY value
is used after it's been overwritten.

(cherry picked from commit 1155e7fdc0613ed9e26955a15e1e9fba6056c720)